% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.volume.R
\name{isr.volume}
\alias{isr.volume}
\title{Serum Volume Calculator}
\usage{
isr.volume(subject.sex = c("m", "f"), subject.weight, subject.height)
}
\arguments{
\item{subject.sex}{String for patient sex, \code{"m"} or \code{"f"}.}

\item{subject.weight}{Numeric for subject weight in kilograms.}

\item{subject.height}{Numeric for subject height in centimeters.}
}
\value{
Numeric value.
}
\description{
Calculates the estimated serum volume (mL) based on patient sex, weight, and
height for use in \code{\link[=isr.deconv]{isr.deconv()}}.
}
\examples{
isr.volume("m", 86.2, 181.5)

}
\seealso{
\code{\link[=isr.deconv]{isr.deconv()}}
}
